requirements.txt
# Основные библиотеки
flask==2.3.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.2.2
joblib==1.3.1

# Обработка текста
nltk==3.8.1
# pymystem3==0.2.0  # Заменяем на pymorphy2
# pymorphy2==0.9.1
tqdm==4.66.1  # Для отображения прогресса

# Модели машинного обучения
lightgbm==3.3.5
catboost==1.2.0

# Flask расширения
flask-cors==4.0.0
flask-caching==2.0.2

# Инструменты мониторинга и логирования
prometheus-client==0.17.1
python-json-logger==2.0.7
python-dotenv==1.0.0